requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2==2.9.6
drf_yasg==1.21.5